Extract tell logic to _tell.py
